Fix C2079 fmt enum-format build error in steam_network.cpp
Reported by user Evan in issue #13 thread. spdlog 1.15 bundles fmtlib
v11, which no longer auto-formats enum types — callers must cast to
the underlying integer explicitly or provide a formatter specialization.

Two sites in the Steam connection status callback were passing raw
Steamworks enums through the `{}` format specifier:

  * on_connection_status_changed line 29: `info->m_info.m_eState`
    (ESteamNetworkingConnectionState)
  * on_connection_status_changed line 65: `info->m_info.m_eEndReason`

Both now cast to `int`, matching the existing convention on line 175
(`static_cast<int>(result)`) and in session.cpp:21 / :57.

The error manifested as:

  base.h(2673,45): error C2079: '_' uses undefined struct
  'fmt::v11::detail::type_is_unformattable_for<ESteamNetworkingConnectionState, ...>'

No other spdlog calls in the codebase format raw enums — grepped and
confirmed.

+++ src/network/steam_network.cpp
@@ -26,7 +26,7 @@ struct SteamNetworkTransport::Impl {
 
 // Connection status change callback - uses g_active_transport to update state
 static void on_connection_status_changed(SteamNetConnectionStatusChangedCallback_t* info) {
-    spdlog::info("Steam: connection status changed to {}", info->m_info.m_eState);
+    spdlog::info("Steam: connection status changed to {}", static_cast<int>(info->m_info.m_eState));
 
     if (!g_active_transport || !g_active_transport->impl_) return;
     auto* impl = g_active_transport->impl_.get();
@@ -62,7 +62,7 @@ static void on_connection_status_changed(SteamNetConnectionStatusChangedCallback
 
         case k_ESteamNetworkingConnectionState_ClosedByPeer:
         case k_ESteamNetworkingConnectionState_ProblemDetectedLocally:
-            spdlog::warn("Steam: connection lost (reason: {})", info->m_info.m_eEndReason);
+            spdlog::warn("Steam: connection lost (reason: {})", static_cast<int>(info->m_info.m_eEndReason));
             if (info->m_hConn == impl->connection) {
                 impl->sockets->CloseConnection(info->m_hConn, 0, nullptr, false);
                 impl->connection = k_HSteamNetConnection_Invalid;
